Bundle required ansible collections to be used as a last resort
Rather than requiring additional setup steps, when searching for
ansible collections we will respect whatever collection versions are
available on the host, though in cases where a required collection is
not present we will fall back to the included collections folder. This
is going to create a _ton_ of churn in git commits, and I recognize
this is rather dumb, though I have no way of implicitly demanding a
collection is installed. I can trim the total churn down with some
ignore rules, so hopefully this isn't too gross for the time being.