Improve how we handle logging to stdout or file
I thought it would be neat to dump logs to a local file but ultimately
that became annoying to deal with. Rather than cluttering up every
directory we run from, instead try using a canonical log file location
and simply don't log to a file if we hit any exceptions when doing so.