playbook: Delete various site-*.yml files for unimplemented things
I want to cut down on unimplemented scaffolding, so a good first step
would be to not leave some of these plays around since this implies
that they _could_ be used. Most of these barely had any thought put
into them thus far, so let's not make it easy to use the roles they
refer to until they're actually done.