playbook: syspatch: Catch return code 2 when no patch was installed
Somehow I overlooked this in the man page, or it is a recent change:

    EXIT STATUS
        The syspatch utility exits 0 on success, and >0 if an error
        occurs.  In particular, 2 indicates that applying patches was
        requested but no additional patch was installed.

This commit corrects our syspatch ansible module so that rc==2 is
caught and handled as a successful condition. If there are no patches
to apply, that's a good thing in our case. I'll leave the original
logic I used to determine this until 6.9 becomes -stable, then we'll
simplify this code to only check for return codes.