mail: finally add scaffolding for a simple, half-decent
I decided to not make this complicated and require that all mail
accounts are tied to system users, and not bother with valiases and
what not because OpenSMTPD already handles everything but IMAP. I like
the idea of being able to create a user account that can be passed to
someone who intrinsically has a ~/Maildir and IMAP/POP access.

Next is buttoning up the automation around this core configuration and
seeing if there are opportunities for hardening.