Build our own OpenBSD Vagrant image with Packer
The generic/openbsd7 image I was using before had bash, sudo, and a
few other things installed. This commit replaces it with a custom
OpenBSD image built with Packer that only uses the base installation
media and no supplemental packages.

I didn't document the caveats, but since it's an optional flow I'll
worry about that later. Or never!